requirements.txt
boto3>=1.28.43
Brotli>=1.1.0
click>=8.1.7
fast-langdetect>=0.2.3
loguru>=0.6.0
numpy>=1.21.6,<2.0.0
pydantic>=2.7.2
PyMuPDF>=1.24.9,<=1.24.14
scikit-learn>=1.0.2
torch>=2.2.2
transformers
pdfminer.six==20231228
pycocotools>=2.0.6
torch==2.5.1
torchvision==0.20.1,
transformers==4.50.0,
lmdeploy==0.8.0,
qwen_vl_utils==0.0.10,
matplotlib
doclayout_yolo==0.0.2b1
PyYAML
